COM-550 add creationFundingModal component
